+++ src/hartreeFock.cpp
@@ -160,7 +160,7 @@ int main(int argc, char *argv[]) {
     for (auto &psi : wf.orbitals)
       of << "\"" << psi.symbol(true) << "\" ";
     of << "\n";
-    for (int i = 0; i < wf.rgrid.ngp; i++) {
+    for (size_t i = 0; i < wf.rgrid.ngp; i++) {
       of << wf.rgrid.r[i] << " ";
       for (auto &psi : wf.orbitals)
         of << psi.f[i] << " ";
